requirements.txt
FreeSimpleGUI==5.2.0.post1
pillow==11.3.0
